Check for HOME in iOS
