+++ lib/android.js
@@ -2,28 +2,11 @@ import { DoctorCheck } from './doctor';
 import { ok, nok } from './utils';
 import { fs, system } from 'appium-support';
 import path from 'path';
+import EnvVarAndPathCheck from './env';
+
+
 let checks = [];
 
-// Check env variables
-class EnvVarAndPathCheck extends DoctorCheck {
-  constructor (varName) {
-    super();
-    this.varName = varName;
-  }
-
-  async diagnose () {
-    let varValue = process.env[this.varName];
-    if (typeof varValue === 'undefined') {
-      return nok(`${this.varName} is NOT set!`);
-    }
-    return await fs.exists(varValue) ? ok(`${this.varName} is set to: ${varValue}`) :
-       nok(`${this.varName} is set to \'${varValue}\' but this is NOT a valid path!`);
-  }
-
-  fix () {
-    return `Manually configure ${this.varName}.`;
-  }
-}
 
 checks.push(new EnvVarAndPathCheck('ANDROID_HOME'));
 checks.push(new EnvVarAndPathCheck('JAVA_HOME'));
@@ -35,7 +18,7 @@ class AndroidToolCheck extends DoctorCheck {
     this.toolName = toolName;
     this.toolPath = toolPath;
   }
-  
+
   async diagnose () {
     if (typeof process.env.ANDROID_HOME === 'undefined') {
       return nok(`${this.toolName} could not be found because ANDROID_HOME is NOT set!`);
