+++ lib/ios.js
@@ -6,6 +6,8 @@ import log from './logger';
 import { FixSkippedError } from './doctor';
 import NodeDetector from './node-detector';
 import { fixIt } from './prompt';
+import EnvVarAndPathCheck from './env';
+
 
 let checks = [];
 let fixes = {};
@@ -49,7 +51,7 @@ class XcodeCmdLineToolsCheck extends DoctorCheck {
     return stdout.match(/install-time/) ? ok('Xcode Command Line Tools are installed.') :
       nok(errMess);
   }
-  
+
   async fix () {
     log.info(`The following command need be executed: xcode-select --install`);
     let yesno = await fixIt();
@@ -150,6 +152,8 @@ class NodeBinaryCheck extends DoctorCheck {
   }
 }
 checks.push(new NodeBinaryCheck());
+
+checks.push(new EnvVarAndPathCheck('HOME'));
 
 export { fixes, XcodeCheck, XcodeCmdLineToolsCheck, DevToolsSecurityCheck,
          AuthorizationDbCheck, NodeBinaryCheck };
